# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "A Go service for investigating \"Deep inspection\" of VMs in VMware vSphere",
        "title": "VM Deep Inspection Demo API",
        "contact": {},
        "version": "0.1"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/vms/inspect-snapshot": {
            "post": {
                "description": "Run virt-inspector or virt-v2v-inspector on a VM snapshot using VDDK",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vms"
                ],
                "summary": "Inspect a VM snapshot directly",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"web-server-01\"",
                        "description": "Original VM name",
                        "name": "vms",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "\"inspection-snapshot\"",
                        "description": "Snapshot name",
                        "name": "snapshot",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "\"virt-inspector\"",
                        "description": "Inspector type: 'virt-inspector' (default) or 'virt-v2v-inspector'",
                        "name": "inspector",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Inspection completed successfully",
                        "schema": {
                            "$ref": "#/definitions/types.VMInspectionResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "VM or snapshot not found",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "types.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "VALIDATION_ERROR"
                },
                "details": {
                    "type": "string",
                    "example": "VM ID is required"
                },
                "error": {
                    "type": "string",
                    "example": "Invalid request"
                }
            }
        },
...
        "types.VMInspectionResponse": {
            "type": "object",
            "properties": {
                "inspector_type": {
                    "type": "string",
                    "example": "virt-inspector"
                },
                "message": {
                    "type": "string",
                    "example": "Inspection completed successfully"
                },
                "snapshot_name": {
                    "type": "string",
                    "example": "backup-snapshot"
                },
                "status": {
                    "type": "string",
                    "example": "completed"
                },
                "virt_inspector": {},
                "virt_v2v": {},
                "vm_name": {
                    "type": "string",
                    "example": "web-server-01"
                }
            }
        },
...
    }
}